+++ src/Hyyan/WPI/Pages.php
@@ -27,21 +27,32 @@ class Pages
     {
 
         $method = array($this, 'getPostTranslationID');
+        $pages = apply_filters(HooksInterface::PAGES_LIST, array(
+            'shop',
+            'cart',
+            'checkout',
+            'terms',
+            'myaccount',
+        ));
 
-        // shop page
-        add_filter('woocommerce_get_shop_page_id', $method);
+        foreach ($pages as $page) {
+            add_filter(sprintf('woocommerce_get_%s_page_id', $page), $method);
+            add_filter(sprintf('option_woocommerce_%s_page_id', $page), $method);
+        }
 
-        //cart page
-        add_filter('woocommerce_get_cart_page_id', $method);
+        /* To generate the correct url for shop page */
+        add_filter(
+                'pll_get_archive_url'
+                , array($this, 'translateShopUrl')
+                , 10
+                , 2
+        );
 
-        // checkout page
-        add_filter('woocommerce_get_checkout_page_id', $method);
+        if (!is_admin()) {
 
-        // terms page (Categories,Tags)
-        add_filter('woocommerce_get_terms_page_id', $method);
-
-        // myaccount page
-        add_filter('woocommerce_get_myaccount_page_id', $method);
+            /* To get product from current language in the shop page */
+            add_filter('parse_request', array($this, 'correctShopPage'));
+        }
     }
 
     /**
@@ -60,6 +71,74 @@ class Pages
         }
 
         return $id;
+    }
+
+    /**
+     * Correct the shop page to display products from currrent language only
+     *
+     * @param  \WP     $wp wordpress instance
+     * @return boolean false if the current language is the same as default
+     *                    language or if the "pagename" var is empty
+     */
+    public function correctShopPage(\WP $wp)
+    {
+        if (pll_default_language() === pll_current_language()) {
+            return false;
+        }
+
+        if (empty($wp->query_vars['pagename'])) {
+            return false;
+        }
+
+        $shopPage = get_post(wc_get_page_id('shop'));
+
+        /* Explode by / for children page */
+        $page = explode('/', $wp->query_vars['pagename']);
+
+        if (
+                isset($shopPage->post_name) &&
+                $shopPage->post_name == $page[count($page) - 1]
+        ) {
+            unset($wp->query_vars['page']);
+            unset($wp->query_vars['pagename']);
+            $wp->query_vars['post_type'] = 'product';
+        }
+    }
+
+    /**
+     * Translate the shop page name in the given shop url
+     *
+     * @param string $url      complete url
+     * @param string $language the current language
+     *
+     * @return string translated url
+     */
+    public function translateShopUrl($url, $language)
+    {
+        $result = $url;
+
+        if (!is_post_type_archive('product')) {
+            return $result;
+        }
+
+        $shopPageID = get_option('woocommerce_shop_page_id');
+        $shopPage = get_post($shopPageID);
+
+        if ($shopPage) {
+
+            $shopPageTranslatedID = pll_get_post($shopPageID, $language);
+            $shopPageTranslation = get_post($shopPageTranslatedID);
+
+            if ($shopPageTranslation) {
+                $result = str_replace(
+                        $shopPage->post_name
+                        , $shopPageTranslation->post_name
+                        , $url
+                );
+            }
+        }
+
+        return $result;
     }
 
 }